fix: troca tinyint por integer

+++ database/migrations/2026_01_24_000006_create_avaliacoes_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('avaliacoes', function (Blueprint $table) {
             $table->id();
-            $table->tinyint('nota');
+            $table->integer('nota');
             $table->foreignId('user_id');
             $table->foreignId('produtos_id');
             $table->timestamps();
